requirements.txt
openwakeword==0.6.0
speechbrain==1.0.2
datasets==3.2.0
scipy==1.12.0
matplotlib==3.10.0
ipython==8.30.0
acoustics==0.2.6
mutagen==1.47.0
audiomentations==0.38.0
torch-audiomentations==0.11.1
pronouncing==0.2.0
onnx==1.17.0
onnxruntime==1.20.1
speexdsp-ns==0.1.2
deep-phonemizer==0.0.19
